Remove secret from the source tree

+++ generate-secret.sh
@@ -1,8 +1,25 @@
 #!/usr/bin/env bash
 
-KAFKA_BROKERS=test
-CLIENT_ID=srvc-acct-750d7986-9f79-48f7-ae66-21b5106716a9
-CLIENT_SECRET=4d445d69-ed01-4c5f-b4bb-cc2717e09d15
+VALID="true"
+
+if [[ -z "${KAFKA_BROKERS}" ]]; then
+  echo "Must set KAFKA_BROKERS environment variable"
+  VALID="false"
+fi
+
+if [[ -z "${CLIENT_ID}" ]]; then
+  echo "Must set CLIENT_ID environment variable"
+  VALID="false"
+fi
+
+if [[ -z "${CLIENT_SECRET}" ]]; then
+  echo "Must set CLIENT_SECRET environment variable"
+  VALID="false"
+fi
+
+if [[ "${VALID}" == "false" ]]; then
+  exit 1
+fi
 
 kubectl delete secret camel-quarkus-kafka-demo-secret
 
@@ -12,10 +29,3 @@ kubectl create secret generic camel-quarkus-kafka-demo-secret \
   --from-literal=OAUTH_CLIENT_SECRET="${CLIENT_SECRET}"
 
 kubectl label secret camel-quarkus-kafka-demo-secret "app.kubernetes.io/name"="camel-quarkus-kafka-demo"
-
-kubectl get secret camel-quarkus-kafka-demo-secret -o yaml > src/main/kubernetes/openshift.yml
-
-sed -i "/creationTimestamp/d" src/main/kubernetes/openshift.yml
-sed -i "/namespace/d" src/main/kubernetes/openshift.yml
-sed -i "/resourceVersion/d" src/main/kubernetes/openshift.yml
-sed -i "/uid/d" src/main/kubernetes/openshift.yml
